# src/docs/swagger.yml
openapi: 3.0.0
info:
  title: UnaHur Anti-Social Con MongoDb
  description: Documentación para la red UnaHur Anti-Social
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  schemas:
...
    Comment:
      type: object
      properties:
        _id:
          type: string
          example: "60c72b2f9b1d8e3a5c8d9876"
        postId:
          type: string
          example: "60c72b2f9b1d8e3a5c8d2222"
        userId:
          type: object
          properties:
            _id:
              type: string
              example: "60c72b2f9b1d8e3a5c8d1234"
            userName:
              type: string
              example: "usuario123"
        text:
          type: string
          example: "Este es un comentario"

        createdAt:
          type: string
          format: date-time
          example: "2024-06-01T12:34:56.789Z"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Descripción del error"
    ErrorServer:
      type: object
      properties:
        error:
          type: string
          example: "Error inesperado en el servidor"
paths:
...
  /comments:
    get:
      summary: Obtener todos los comentarios
      tags:
        - Comments
      responses:
        "200":
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "500":
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorServer"

    post:
      summary: Crear un nuevo comentario
      tags:
        - Comments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                postId:
                  type: string
                  example: "60c72b2f9b1d8e3a5c8d1234"
                userId:
                  type: string
                  example: "60c72b2f9b1d8e3a5c8d5678"
                text:
                  type: string
                  example: "Este es un comentario de prueba"
                createdAt:
                  type: string
                  format: date-time
                  example: "2024-06-01T15:30:00Z"
              required:
                - postId
                - userId
                - text
      responses:
        "201":
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Datos inválidos o faltantes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorServer"
...